Fold /draft-capital into /league as default tab + mobile dropdown nav
The public landing page "League" button used to drop unauthenticated
mobile visitors on /draft-capital with no way to reach the rest of
the public /league features.  Draft Capital is genuinely the most
valuable public page for a dynasty audience, so this change keeps
it as the landing experience AND threads it into the /league tab
rail so every other public section (History, Rivalries, Awards,
Records, Franchises, Trades, Draft, Weekly, Superlatives, Archives,
Home/Overview) stays one click away.

Changes:
- frontend/app/league/sections/draft-capital.jsx — new DraftCapitalSection
  component holding the auction-dollar board, per-team bars, pick-value
  grid, and per-round picks.  Same /api/draft-capital data source as
  before; no auth required.
- frontend/app/league/LeagueClient.jsx — "Draft Capital" added as the
  FIRST tab and the default selection.  DEFAULT_TAB constant replaces
  the hardcoded "overview" so ``?tab=`` is omitted from the URL only
  for the default tab (clean shareable link).
- frontend/app/league/LeagueClient.jsx — mobile-only ``<select>``
  dropdown mirrors the SubNav so all 12 sections stay reachable on
  small screens without a cramped horizontal scroll.  Desktop still
  sees the horizontal SubNav.
- frontend/app/page.jsx — landing-page "League" button now points at
  /league (was /draft-capital).  Mobile users now land on the Draft
  Capital tab inside the /league shell with the rest of the public
  features in the dropdown.
- frontend/app/draft-capital/page.jsx — reduced to a server-side
  redirect → /league?tab=draft-capital so old bookmarks, links from
  deployed clients, and any social previews continue to resolve.

Tests / build:
- backend pytest: 154 passed
- frontend vitest: 384 passed
- next build: clean (/league now 16.2 kB; /draft-capital shrinks to
  a 182 B redirect stub)
