+++ frontend/app/page.jsx
@@ -45,7 +45,7 @@ function LandingHome() {
         <h1 style={{ margin: "0 0 8px", fontSize: "1.4rem" }}>Risk It To Get The Brisket</h1>
         <p className="muted" style={{ marginBottom: "var(--space-lg)" }}>Choose where you want to go.</p>
         <div className="grid-responsive" style={{ display: "grid", gridTemplateColumns: "1fr 1fr", gap: 12 }}>
-          <Link href="/draft-capital" className="button" style={{ textAlign: "center", padding: "14px 12px" }}>
+          <Link href="/league" className="button" style={{ textAlign: "center", padding: "14px 12px" }}>
             League
           </Link>
           <Link href="/login" className="button button-primary" style={{ textAlign: "center", padding: "14px 12px" }}>
